Make `expr2csgrammar` use `add_rule!`
Previously, it (almost) replicated the behavior in `add_rule!`, but the check
for duplicate rules was not present. This solution makes sure that the logic
for adding rules to a grammar is consistent across different approaches to
creating a grammar.

There might be a performance hit here, given that `add_rule!` recalculates
some of the fields of the grammar struct for each rule. This is not necessary
in the case of the macro where those recalculations can be done after all
rules have been added. For now, I don't think this is a problem.